pyproject.toml
[project]
name = "todolist"
version = "0.1.0"
description = ""
authors = [
    {name = "sufiks",email = "sufx@rambler.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "python-decouple (>=3.8,<4.0)",
    "poetry (==2.1.1)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "social-auth-app-django (>=5.4.3,<6.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"


[tools.isort]
py_version = 3.13
profile = 'django'
line_lenght = 120
multi_line_output = 2
known_first_party = ["todolist", "core"]
